requirements.txt
future==0.17.1
llvmlite==0.29.0
numba==0.44.1
numpy==1.16.4
Pillow==6.1.0
pkg-resources==0.0.0
pygame==1.9.6
pyglet==1.3.2
pyserial==3.4
